Move everything RelayRace related to a new file
Also add callbacks instead of hardcoding normal replay so it can be used
with shared libraries.
